+++ buildcontext/git.go
@@ -34,6 +34,7 @@ const (
 type gitResolver struct {
 	cleanCollection   *cleanup.Collection
 	gitBranchOverride string
+	lfsInclude        string
 	gitImage          string
 	projectCache      *synccache.SyncCache // "gitURL#gitRef" -> *resolvedGitProject
 	buildFileCache    *synccache.SyncCache // project ref -> local path
@@ -193,6 +194,11 @@ func (gr *gitResolver) resolveGitProject(ctx context.Context, gwClient gwclient.
 		if len(keyScans) > 0 {
 			gitOpts = append(gitOpts, llb.KnownSSHHosts(strings.Join(keyScans, "\n")))
 		}
+		if gr.lfsInclude != "" {
+			// TODO this should eventually be infered by the contents of a COPY command, which means the call to resolveGitProject will need to be lazy-evaluated
+			// However this makes it really difficult for an Earthfile which first has an ARG EARTHLY_GIT_HASH, then a RUN, then a COPY
+			gitOpts = append(gitOpts, llb.LFSInclude(gr.lfsInclude))
+		}
 
 		gitState := llb.Git(gitURL, gitRef, gitOpts...)
 		gitImage := gr.gitImage
@@ -325,6 +331,9 @@ func (gr *gitResolver) resolveGitProject(ctx context.Context, gwClient gwclient.
 		}
 		if len(keyScans) > 0 {
 			gitOpts = append(gitOpts, llb.KnownSSHHosts(strings.Join(keyScans, "\n")))
+		}
+		if gr.lfsInclude != "" {
+			gitOpts = append(gitOpts, llb.LFSInclude(gr.lfsInclude))
 		}
 
 		rgp := &resolvedGitProject{