+++ features/features_test.go
@@ -117,6 +117,7 @@ func TestAvailableFlags(t *testing.T) {
 		{"try", "TryFinally"},
 		{"no-network", "NoNetwork"},
 		{"arg-scope-and-set", "ArgScopeSet"},
+		{"earthly-ci-runner-arg", "EarthlyCIRunnerArg"},
 	} {
 		tt := tt
 		t.Run(tt.flag, func(t *testing.T) {