+++ CHANGELOG.md
@@ -11,6 +11,9 @@ All notable changes to [Earthly](https://github.com/earthly/earthly) will be doc
 ### Added
 - Experimental support for performing a `git lfs pull --include=<path>` when referencing a remote target on the cli, when used with the new `--git-lfs-pull-include` flag. [#2992](https://github.com/earthly/earthly/pull/2922)
 
+### Fixed
+- `SAVE IMAGE <img>` was incorrectly pushed when earthly was run with the `--push` cli flag (this restores the requirement that images that are pushed must be defined with `SAVE IMAGE --push <img>`).  [#2923](https://github.com/earthly/earthly/issues/2923)
+
 ## v0.7.4 - 2023-04-12
 
 ### Changed